Update to QWERTY Layout
I'm not sure I'm getting much value out of using Canary in addition to the ortho-split lifestyle, and it makes it much harder when I occasionally have to use a qwerty keyboard (and vim motions are in weird places). Updating to Qwerty for my base layer.

Additionally moved files from jbowe-tri-layer to jbowe, simplifying keymap name.